chore: move warp_image, remove useless tests
