+++ src/lane_assist/helpers.py
@@ -4,10 +4,9 @@ import cv2
 import numpy as np
 
 from config import config
-from lane_assist.preprocessing.birdview import warp_image
 from lane_assist.preprocessing.calibrate import CameraCalibrator
 from lane_assist.preprocessing.gamma import adjust_gamma
-from lane_assist.preprocessing.stitching import stitch_images
+from lane_assist.preprocessing.stitching import stitch_images, warp_image
 from utils.video_stream import VideoStream
