--- tests/test_line_detection/test_data.py
@@ -1,21 +1,0 @@
-import numpy as np
-from lane_assist.line_detection.line import Line, LineType
-
-straight = [
-	Line(np.array([[316, 792], [315, 785], [315, 778], [315, 771], [315, 764], [315, 757], [315, 750], [315, 743], [316, 736], [316, 729], [316, 722], [316, 715], [317, 708], [317, 701], [317, 694], [317, 687], [317, 680], [317, 673], [317, 666], [317, 659], [317, 652], [318, 645], [318, 638], [318, 631], [318, 624], [318, 617], [318, 610], [319, 603], [319, 596], [319, 589], [319, 582], [319, 575], [319, 568], [320, 561], [320, 554], [320, 547], [320, 540], [320, 533], [321, 526], [321, 519], [321, 512], [321, 505], [322, 498], [322, 491], [322, 484], [322, 477], [323, 470], [323, 463], [323, 456], [323, 449], [324, 442], [324, 435], [324, 428], [324, 421], [325, 414], [325, 407], [325, 400], [326, 393], [326, 386], [326, 379], [326, 372], [327, 365], [327, 358], [327, 351], [327, 344], [327, 337], [327, 330], [328, 323], [327, 316], [328, 309]]), line_type=LineType.SOLID),
-	Line(np.array([[467, 792], [467, 785], [468, 778], [468, 771], [469, 764], [469, 757], [470, 750], [470, 743], [470, 736], [470, 729], [471, 722], [471, 715], [471, 708], [472, 701], [472, 694], [473, 687], [473, 680], [473, 673], [474, 666], [474, 659], [475, 652], [475, 645], [475, 638], [476, 631], [476, 624], [476, 617], [477, 610], [477, 603], [478, 596], [478, 589], [478, 582], [479, 575], [479, 568], [479, 561], [480, 554], [480, 547], [481, 540], [481, 533], [481, 526], [482, 519], [482, 512], [482, 505], [483, 498], [483, 491], [484, 484], [484, 477], [484, 470], [485, 463], [485, 456], [485, 449], [486, 442], [486, 435], [487, 428], [487, 421], [487, 414], [488, 407], [488, 400], [488, 393], [489, 386], [490, 379], [490, 372], [490, 365], [491, 358], [491, 351], [491, 344], [492, 337], [492, 330], [492, 323], [493, 316], [493, 309]]), line_type=LineType.SOLID),
-]
-corner = [
-	Line(np.array([[324, 792], [322, 785], [318, 778], [314, 771], [308, 764], [302, 757], [294, 750], [285, 743]]), line_type=LineType.SOLID),
-	Line(np.array([[480, 792], [480, 785], [479, 778], [477, 771], [476, 764], [474, 757], [472, 750], [470, 743], [468, 736], [465, 729], [462, 722], [457, 715], [453, 708], [448, 701], [442, 694], [436, 687], [431, 680], [425, 673], [418, 666], [410, 659], [402, 652], [393, 645], [384, 638], [374, 631], [364, 624], [353, 617], [342, 610]]), line_type=LineType.SOLID),
-]
-crossing = [
-	Line(np.array([[108, 449], [105, 442], [102, 435], [101, 428], [99, 421], [99, 414], [97, 407], [97, 400], [95, 393], [93, 386], [92, 379], [91, 372], [91, 365], [89, 358], [87, 351], [88, 344], [85, 337], [84, 330], [84, 323], [82, 316], [83, 309], [80, 302], [79, 295], [77, 288], [74, 281], [73, 274], [74, 267], [74, 260], [73, 253], [72, 246], [71, 239], [68, 232], [69, 225], [66, 218], [67, 211], [65, 204], [64, 197], [66, 190], [63, 183], [61, 176], [60, 169], [56, 162], [59, 155], [58, 148], [55, 141], [57, 134], [55, 127], [52, 120], [52, 113], [53, 106], [49, 99], [46, 92], [49, 85], [46, 78], [42, 71], [40, 64], [44, 57], [41, 50], [38, 43], [42, 36], [41, 29]]), line_type=LineType.SOLID),
-	Line(np.array([[331, 785], [328, 757], [328, 750], [327, 743], [327, 736], [327, 729], [326, 722], [326, 715], [296, 379], [295, 372], [294, 365], [293, 358], [292, 351], [292, 344], [291, 337], [290, 330], [290, 323], [289, 316], [287, 274], [286, 267], [286, 260], [285, 253], [284, 246], [283, 239], [282, 232], [282, 225], [281, 218], [280, 211], [278, 155], [277, 148], [276, 141], [276, 134], [275, 127], [274, 120], [273, 113], [273, 106], [272, 99], [271, 92], [269, 36], [269, 29], [267, 22]]), line_type=LineType.DASHED),
-	Line(np.array([[482, 792], [483, 785], [483, 778], [483, 771], [482, 764], [482, 757], [482, 750], [482, 743], [482, 736], [480, 729], [481, 722], [480, 715], [481, 708], [480, 701], [480, 694], [480, 687], [480, 680], [474, 470], [474, 463], [474, 456], [474, 449], [474, 442], [473, 435], [473, 428], [473, 421], [472, 414], [472, 407], [472, 400], [472, 393], [472, 386], [472, 379], [471, 372], [471, 365], [471, 358], [470, 351], [469, 344], [470, 337], [470, 330], [470, 323], [470, 316], [469, 309], [469, 302], [469, 295], [469, 288], [469, 281], [468, 274], [468, 267], [468, 260], [468, 253], [468, 246], [468, 239], [468, 232], [467, 225], [468, 218], [467, 211], [467, 204], [467, 197], [466, 190], [466, 183], [465, 176], [465, 169], [466, 162], [464, 155], [464, 148], [464, 141], [463, 134], [463, 127], [464, 120], [463, 113], [462, 106], [463, 99], [464, 92], [462, 85], [462, 78], [463, 71], [463, 64], [461, 57], [462, 50], [462, 43], [461, 36], [460, 29], [460, 22]]), line_type=LineType.SOLID),
-	Line(np.array([[717, 463], [716, 456], [718, 449], [719, 442], [720, 435], [720, 428], [722, 421], [723, 414], [723, 407], [723, 400], [719, 393], [721, 386], [728, 379], [726, 365], [730, 358], [728, 337], [731, 330], [734, 323], [731, 211], [735, 204], [733, 197], [734, 190], [738, 183], [739, 176]]), line_type=LineType.SOLID),
-]
-stopline = [
-	Line(np.array([[316, 792], [315, 785], [315, 778], [315, 771], [315, 764], [318, 757], [315, 743], [315, 736], [316, 729], [316, 722], [316, 715], [316, 708], [317, 701], [317, 694], [317, 687], [317, 680], [317, 673], [317, 666], [317, 659], [317, 652], [318, 645], [318, 638], [318, 631], [318, 624], [318, 617], [318, 610], [318, 603], [319, 596], [319, 589], [319, 582], [319, 575], [319, 568], [320, 561], [320, 554], [320, 547], [320, 540], [320, 533], [320, 526], [321, 519], [321, 512], [321, 505], [321, 498], [322, 491], [322, 484], [322, 477], [322, 470], [323, 463], [323, 456], [323, 449], [323, 442], [324, 435], [324, 428], [324, 421], [324, 414], [325, 407], [325, 400], [325, 393], [325, 386], [325, 379], [326, 372], [326, 365], [326, 358], [326, 351], [327, 344], [327, 337], [327, 330], [328, 323], [327, 316], [328, 309], [328, 302], [328, 295], [329, 288]]), line_type=LineType.SOLID),
-	Line(np.array([[467, 792], [467, 785], [468, 778], [468, 771], [469, 764], [469, 757], [470, 743], [470, 736], [470, 729], [471, 722], [471, 715], [471, 708], [472, 701], [472, 694], [473, 687], [473, 680], [474, 673], [474, 666], [474, 659], [475, 652], [475, 645], [475, 638], [476, 631], [476, 624], [477, 617], [477, 610], [478, 603], [478, 596], [478, 589], [479, 582], [479, 575], [479, 568], [480, 561], [480, 554], [480, 547], [481, 540], [481, 533], [482, 526], [482, 519], [482, 512], [482, 505], [483, 498], [483, 491], [484, 484], [484, 477], [485, 470], [485, 463], [485, 456], [486, 449], [486, 442], [487, 435], [487, 428], [487, 421], [488, 414], [488, 407], [488, 400], [489, 393], [489, 386], [490, 379], [490, 372], [490, 365], [491, 358], [492, 351], [491, 344], [492, 337], [492, 330], [492, 323], [493, 316], [493, 309], [494, 302], [495, 295]]), line_type=LineType.SOLID),
-]
