--- tests/test_line_generation/test_data.py
@@ -1,7 +1,0 @@
-import numpy as np
-
-STRAIGHT = np.array([[391.1740172490591, 791.9999999999994], [391.1987852135822, 790.2688172043006], [391.22612039566917, 788.5376344086018], [391.2559437004717, 786.806451612903], [391.2881760331413, 785.075268817204], [391.3227382988298, 783.3440860215052], [391.35955140268874, 781.6129032258063], [391.39853624986984, 779.8817204301074], [391.4396137455247, 778.1505376344085], [391.482704794805, 776.4193548387096], [391.5277303028624, 774.6881720430107], [391.5746111748485, 772.9569892473119], [391.623268315915, 771.2258064516129], [391.67362263121356, 769.494623655914], [391.7255950258958, 767.763440860215], [391.7791064051134, 766.0322580645161], [391.834077674018, 764.3010752688172], [391.8904297377612, 762.5698924731182], [391.94808350149475, 760.8387096774193], [392.0069598703702, 759.1075268817203], [392.0669797495393, 757.3763440860214], [392.1280640441537, 755.6451612903224], [392.19013365936496, 753.9139784946235], [392.25310950032485, 752.1827956989246], [392.31691247218487, 750.4516129032256], [392.3814634800512, 748.7204301074397], [392.4507709352304, 746.989247311741], [392.51733097373153, 745.2580645160422], [392.58188400127244, 743.5268817203437], [392.64524568191547, 741.7956989246447], [392.70593824865614, 740.0645161289461], [392.7701235088831, 738.3333333332473], [392.8367472368893, 736.6021505375486], [392.90404926421263, 734.8709677418499], [392.97270910794316, 733.1397849461513], [393.0397467568433, 731.4086021504525], [393.10596119149113, 729.6774193547539], [393.1731735907539, 727.9462365590551], [393.2405911846936, 726.2150537633564], [393.3109064684455, 724.4838709676576], [393.3844948812908, 722.7526881719588], [393.4613302411194, 721.0215053762602], [393.5414143745902, 719.2903225805613], [393.6214661153109, 717.5591397848626], [393.69937384565793, 715.8279569891639], [393.7722024902016, 714.0967741934653], [393.8399514400555, 712.3655913977665], [393.90572881613065, 710.6344086020679], [393.97164623605056, 708.9032258063692], [394.0408118207262, 707.1720430106703], [394.1132249612716, 705.4408602149716], [394.1859505822564, 703.7096774192731], [394.2568770660575, 701.9784946235742], [394.3227232462824, 700.2473118278756], [394.3834909495901, 698.5161290321768], [394.4419422059296, 696.7849462364779], [394.49975900276394, 695.0537634407792], [394.5576208571831, 693.3225806450805], [394.616207286277, 691.5913978493818], [394.6735838583301, 689.8602150536832], [394.7288383399157, 688.1290322579845], [394.78483748942335, 686.3978494622858], [394.8407885369147, 684.666666666587], [394.89833839800343, 682.9354838708884], [394.95857125560167, 681.2043010751896], [395.01833283521825, 679.4731182794908], [395.0783026539432, 677.7419354837921], [395.136894806568, 676.0107526880932], [395.19262665499826, 674.2795698923946], [395.2487134841304, 672.5483870966958], [395.3043625240263, 670.8172043009971], [395.3608721637922, 669.0860215052984], [395.4197537809551, 667.3548387095998], [395.4809714512246, 665.623655913901], [395.5445172590792, 663.8924731182024], [395.6104112977658, 662.1612903225035], [395.67870166930095, 660.4301075268048], [395.7464570733493, 658.6989247311062], [395.81412175334185, 656.9677419354075], [395.8807458466797, 655.2365591397087], [395.94453399127406, 653.50537634401], [396.0055245469607, 651.7741935483112], [396.0637278648064, 650.0430107526125], [396.12069465639223, 648.3118279569137], [396.17910937955475, 646.5806451612151], [396.23965155138393, 644.8494623655163], [396.3030006889696, 643.1182795698177], [396.36844220337207, 641.3870967741188], [396.4337834950926, 639.6559139784202], [396.5015298876241, 637.9247311827214], [396.5717599272967, 636.1935483870227], [396.6436808441722, 634.4623655913241], [396.71649986831267, 632.7311827956253], [396.7894242297796, 630.9999999999266], [396.8616611586351, 629.2688172042277], [396.9334634644632, 627.537634408529], [397.0071944784756, 625.8064516128304], [397.0805196088929, 624.0752688171316], [397.1528985300293, 622.3440860214329], [397.22413944270585, 620.6129032257342], [397.2916237705819, 618.8817204300354], [397.3574308603077, 617.1505376343367], [397.4221620479744, 615.419354838638], [397.48572161665885, 613.6881720429393], [397.5507568821344, 611.9569892472406], [397.6152172371834, 610.2258064515419], [397.6782138296121, 608.494623655843], [397.74042617651077, 606.7634408601444], [397.802533794969, 605.0322580644456], [397.8652162020767, 603.3010752687469], [397.92915291492363, 601.5698924730483], [397.9946749240925, 599.8387096773496], [398.0587376763998, 598.1075268816509], [398.1229945416535, 596.3763440859522], [398.188569645706, 594.6451612902533], [398.254844481873, 592.9139784945547], [398.32489234277034, 591.182795698856], [398.39708859802226, 589.4516129031572], [398.47002160567445, 587.7204301074586], [398.5443708828168, 585.9892473117596], [398.61680789170435, 584.2580645160609], [398.6887020175836, 582.5268817203622], [398.7615259128142, 580.7956989246635], [398.83448680745806, 579.0645161289648], [398.90679193157706, 577.3333333332661], [398.976359624753, 575.6021505375674], [399.0406606951507, 573.8709677418686], [399.09957762771796, 572.1397849461699], [399.1566244269611, 570.4086021504713], [399.21265377722824, 568.6774193547725], [399.2682319427611, 566.9462365590739], [399.323925187801, 565.2150537633751], [399.38029977658965, 563.4838709676764], [399.438924443742, 561.7526881719775], [399.5024479663585, 560.0215053762789], [399.57075891825, 558.2903225805801], [399.6437738819957, 556.5591397848814], [399.721437448944, 554.8279569891827], [399.80372221921243, 553.096774193484], [399.89062880168717, 551.3655913977852], [399.97904907545706, 549.6344086020865], [400.06615537263195, 547.9032258063879], [400.14641096847953, 546.1720430106891], [400.2204965978622, 544.4408602149904], [400.2890369781044, 542.7096774192917], [400.35374648941945, 540.978494623593], [400.42034111266145, 539.2473118278942], [400.4879148250435, 537.5161290321955], [400.55834981583837, 535.7849462364967], [400.63204661044176, 534.053763440798], [400.7060862080215, 532.3225806450993], [400.781034872819, 530.5913978494007], [400.85484492027075, 528.8602150537018], [400.92649086410177, 527.1290322580031], [400.998726209853, 525.3978494623044], [401.0706449347379, 523.6666666666057], [401.1437807015218, 521.935483870907], [401.2191044402686, 520.2043010752084], [401.2933486236389, 518.4731182795095], [401.367079515874, 516.7419354838108], [401.44086338121565, 515.0107526881121], [401.5152664839053, 513.2795698924134], [401.5908550881845, 511.5483870967147], [401.6681954582947, 509.817204301016], [401.7457626994335, 508.08602150531715], [401.8218189283767, 506.3548387096185], [401.89932418489195, 504.62365591391966], [401.9775467094465, 502.892473118221], [402.05749737504436, 501.1612903225223], [402.14071629864554, 499.43010752682363], [402.2270530853382, 497.6989247311248], [402.31638534898013, 495.96774193542615], [402.4068760796607, 494.23655913972743], [402.49587077674835, 492.50537634402866], [402.57684826817086, 490.7741935483299], [402.64934994078163, 489.0430107526313], [402.7154751127006, 487.3118279569326], [402.77842883953514, 485.5806451612338], [402.8448252094039, 483.84946236553515], [402.9158202538002, 482.11827956983643], [402.9912039069572, 480.3870967741377], [403.0707941118766, 478.6559139784389], [403.15185903936924, 476.92473118274006], [403.234093637408, 475.19354838704146], [403.3168443274589, 473.46236559134263], [403.3976632648938, 471.7311827956439], [403.4788005280767, 469.9999999999452], [403.5603956737429, 468.2688172042465], [403.64154267910624, 466.5376344085477], [403.72133552138007, 464.80645161284906], [403.7965768169244, 463.0752688171504], [403.8658354394239, 461.3440860214517], [403.93061472910694, 459.61290322575303], [403.9972435717578, 457.8817204300542], [404.0669233000095, 456.1505376343555], [404.14088325526365, 454.4193548386568], [404.21898668166153, 452.6881720429581], [404.29563876671955, 450.95698924725934], [404.3726636119589, 449.22580645156074], [404.4507235638764, 447.4946236558619], [404.5294353894465, 445.7634408601632], [404.61147514387534, 444.0322580644644], [404.6948209593775, 442.30107526876577], [404.7782964676574, 440.569892473067], [404.8621194064495, 438.8387096773683], [404.9431319697219, 437.10752688166957], [405.02115575379423, 435.3763440859708], [405.09598434621716, 433.6451612902721], [405.1675575890262, 431.9139784945733], [405.23947911478797, 430.1827956988746], [405.312201934896, 428.4516129031759], [405.3861790607433, 426.72043010747717], [405.4618635037231, 424.98924731177846], [405.5397082752288, 423.25806451607974], [405.62159848718704, 421.526881720381], [405.7095505274066, 419.7956989246822], [405.8033433701675, 418.0645161289836], [405.8989502069375, 416.3333333332848], [405.99403291439666, 414.60215053758617], [406.08594904786156, 412.87096774188745], [406.17446783943177, 411.13978494618874], [406.25933051243777, 409.4086021504899], [406.3402502814409, 407.67741935479125], [406.4169123522338, 405.94623655909254], [406.48897392184034, 404.2150537633938], [406.56303470866266, 402.4838709676951], [406.64203478099836, 400.75268817199645], [406.7291629980796, 399.0215053762976], [406.82487602461697, 397.29032258059885], [406.9230645392193, 395.5591397849002], [407.02167683867344, 393.8279569892015], [407.11980690019294, 392.0967741935028], [407.2165487009911, 390.365591397804], [407.314132956848, 388.6344086021053], [407.4123897884913, 386.9032258064065], [407.5087487215958, 385.1720430107078], [407.6037760204033, 383.4408602150091], [407.6950754738426, 381.70967741931037], [407.7821920003072, 379.97849462361165], [407.86753057883044, 378.24731182791294], [407.950185186626, 376.5161290322142], [408.03203801296655, 374.7849462365155], [408.11348958324777, 373.05376344081674], [408.1916208966374, 371.3225806451181], [408.26699821737697, 369.59139784941937], [408.34018780970814, 367.86021505372065], [408.41204235797903, 366.12903225802194], [408.48607586669596, 364.3978494623231], [408.56217082080667, 362.6666666666244], [408.63779802847614, 360.93548387092574], [408.7107326192326, 359.204301075227], [408.7770610373923, 357.47311827952825], [408.83667551008347, 355.7419354838296], [408.8917056779631, 354.0107526881308], [408.9441499058064, 352.27956989243205], [408.99706205812055, 350.54838709677415], [409.0491356958462, 348.8172043010752], [409.10055425360383, 347.08602150537627], [409.15150116596635, 345.3548387096773], [409.20215986750674, 343.62365591397844], [409.25271379279786, 341.8924731182795], [409.3033463764127, 340.16129032258056], [409.35424105292424, 338.4301075268816], [409.40558125690535, 336.6989247311827], [409.45755042292905, 334.96774193548373], [409.5103319855682, 333.23655913978484], [409.5641093793958, 331.5053763440859], [409.6190660389848, 329.77419354838696], [409.6753853989081, 328.043010752688], [409.7332508937386, 326.3118279569891], [409.7928459580494, 324.58064516129014], [409.85435402641326, 322.8494623655912], [409.91795853340324, 321.11827956989225], [409.98384291359224, 319.3870967741933], [410.05219060155326, 317.65591397849437], [410.1231850318591, 315.9247311827954], [410.19700963908286, 314.1935483870965], [410.2738478577974, 312.46236559139754], [410.35388312257567, 310.73118279569854], [410.43729886799065, 308.9999999999996]])
-CORNER = np.array([[402.3731474766126, 791.9999999999994], [402.17066891216587, 790.8333333333328], [401.9641395345462, 789.6666666666663], [401.7531419362021, 788.4999999999997], [401.53725870958215, 787.333333333333], [401.31607244713484, 786.1666666666664], [401.08916574130876, 784.9999999999999], [400.8561211845525, 783.8333333333333], [400.6165213693145, 782.6666666666666], [400.3699488880434, 781.5], [400.1159863331876, 780.3333333333334], [399.8542162971958, 779.1666666666666], [399.5842213725165, 778.0], [399.3055841515982, 776.8333333333334], [399.0178872268895, 775.6666666666667], [398.7207131908389, 774.5000000000001], [398.41364463589497, 773.3333333333334], [398.0962641545063, 772.1666666666667], [397.7681543391213, 771.0000000000001], [397.4288977821886, 769.8333333333334], [397.0780770761567, 768.6666666666667], [396.71527481347425, 767.5000000000001], [396.3400735865897, 766.3333333333334], [395.95205598795167, 765.1666666666667], [395.5508046100086, 764.0], [395.1359020451632, 762.8333333332446], [394.6998822398713, 761.666666666578], [394.2519210084948, 760.4999999999114], [393.7914806859964, 759.3333333332448], [393.3182340974812, 758.1666666665784], [392.83124604897415, 756.9999999999118], [392.3270210313515, 755.8333333332454], [391.80506942623833, 754.6666666665789], [391.26198815713843, 753.4999999999122], [390.6962531152659, 752.3333333332458], [390.10569693853563, 751.166666666579], [389.4906950772995, 749.9999999999126], [388.85464786253027, 748.8333333332462], [388.1996333123992, 747.6666666665795], [387.52768036884345, 746.4999999999131], [386.84046059814506, 745.3333333332466], [386.1392800687579, 744.16666666658], [385.4205442039145, 742.9999999999135], [384.6828074854422, 741.833333333247], [383.9272106783947, 740.6666666665806], [383.15499201391356, 739.4999999999138], [382.3658304945318, 738.3333333332473], [381.5602946724272, 737.1666666665809], [380.7375265714762, 735.9999999999143], [379.8968787056977, 734.8333333332477], [379.0379140792525, 733.6666666665812], [378.1604061864443, 732.4999999999147], [377.2672986861544, 731.3333333332482], [376.35870757675474, 730.1666666665816], [375.4343342825544, 728.9999999999152], [374.4941204611769, 727.8333333332487], [373.5382480035605, 726.666666666582], [372.5648472081366, 725.4999999999155], [371.5721915981511, 724.3333333332489], [370.56118673880337, 723.1666666665824], [369.534798481957, 721.999999999916], [368.4936035019639, 720.8333333332494], [367.4374445032016, 719.666666666583], [366.36641243113905, 718.4999999999164], [365.2808464723372, 717.3333333332498], [364.1734763659185, 716.1666666665835], [363.04302784574446, 714.9999999999168], [361.8911060135468, 713.8333333332503], [360.7187402118804, 712.6666666665836], [359.527596630377, 711.4999999999172], [358.3165803771063, 710.3333333332506], [357.0856359696121, 709.1666666665841], [355.83363053622116, 707.9999999999176], [354.55778389110236, 706.8333333332511], [353.25793988260153, 705.6666666665844], [351.9269581536268, 704.4999999999999], [350.5737365612917, 703.3333333333331], [349.19824528567887, 702.1666666666665], [347.80045450682996, 700.9999999999998], [346.38033440478677, 699.833333333333], [344.9378551595909, 698.6666666666664], [343.4729869512842, 697.4999999999997], [341.9856999599083, 696.333333333333], [340.4759643655048, 695.1666666666663], [338.9437503481156, 693.9999999999995], [337.38902808778226, 692.8333333333329], [335.8117677645466, 691.6666666666662], [334.2119395584501, 690.4999999999995], [332.58951364953475, 689.3333333333328], [330.94446021784216, 688.1666666666661], [329.27674944341396, 686.9999999999994], [327.5863515062919, 685.8333333333328], [325.8732365865177, 684.6666666666661], [324.13737486413305, 683.4999999999994], [322.37873651917965, 682.3333333333327], [320.5972917316992, 681.1666666666661], [318.79301068173345, 679.9999999999994], [316.96586354932407, 678.8333333333327], [315.1158205145128, 677.6666666666661], [313.24285175734127, 676.4999999999994]])
-CROSSING_LANE_1 = np.array([[218.62088005861526, 616.9999999999997], [218.10984862449948, 613.3711656441716], [217.60647299640567, 609.7423312883434], [217.11047469581905, 606.1134969325152], [216.62157524422474, 602.4846625766871], [216.13949616310802, 598.8558282208589], [215.66395897395404, 595.2269938650307], [215.19468519824798, 591.5981595092026], [214.73139635747503, 587.9693251533744], [214.27381397312044, 584.3404907975462], [213.82165956666933, 580.711656441718], [213.37465465960696, 577.0828220858898], [212.93252077341847, 573.4539877300616], [212.49497942958908, 569.8251533742334], [212.061752149604, 566.1963190184051], [211.63256045494836, 562.5674846625769], [211.20712586710744, 558.9386503067487], [210.78516990756634, 555.3098159509204], [210.36641409781035, 551.6809815950922], [209.95057995932459, 548.0521472392641], [209.53738901359426, 544.4233128834358], [209.1265627821046, 540.7944785276076], [208.71782278634078, 537.1656441717793], [208.31089054778798, 533.5368098159511], [207.9054875879314, 529.9079754601228], [207.5013354282321, 526.2791411042332], [207.122007970892, 522.6503067484053], [206.73722135280693, 519.0214723925776], [206.336224257013, 515.3926380367499], [205.9191254635658, 511.76380368092197], [205.49893094172424, 508.13496932509435], [205.07921011034273, 504.5061349692664], [204.6624789457158, 500.8773006134386], [204.24944780549038, 497.24846625761086], [203.8352496050851, 493.619631901783], [203.4114802449277, 489.9907975459553], [202.98986546809908, 486.3619631901274], [202.57126980897655, 482.7331288342996], [202.1677845116281, 479.10429447847184], [201.77232699131244, 475.475460122644], [201.38450222661092, 471.84662576681615], [200.9931394568243, 468.2177914109884], [200.5900987898109, 464.5889570551606], [200.1824306669999, 460.9601226993328], [199.76591735777532, 457.331288343505], [199.34429481652595, 453.7024539876773], [198.91132260286912, 450.0736196318494], [198.47101222996437, 446.4447852760216], [198.02367705385205, 442.81595092019387], [197.5727336975769, 439.1871165643661], [197.11938076536342, 435.5582822085382], [196.6659682885542, 431.9294478527105], [196.21481565769062, 428.3006134968827], [195.7660361256209, 424.67177914105474], [195.3197429451933, 421.04294478522706], [194.87604936925618, 417.4141104293992], [194.43576463457296, 413.7852760735715], [194.00930300060816, 410.1564417177436], [193.60176848836602, 406.52760736191584], [193.20765214579936, 402.89877300608805], [192.80293881064475, 399.26993865026026], [192.38385070138793, 395.64110429443247], [191.953851062338, 392.01226993860473], [191.53437870110508, 388.3834355827768], [191.12168824003822, 384.7546012269491], [190.72254102182697, 381.12576687112124], [190.32581546912186, 377.4969325152935], [189.92947307407843, 373.86809815946566], [189.52524967039824, 370.23926380363787], [189.11027510608474, 366.6104294478101], [188.69617356142714, 362.9815950919823], [188.2837575821868, 359.35276073615444], [187.8720689919458, 355.7239263803266], [187.45549685029832, 352.0950920244989], [187.0268593199303, 348.4662576686711], [186.59620843357743, 344.83742331284327], [186.16327317474193, 341.2085889570155], [185.73946584940416, 337.5797546011877], [185.32905967107453, 333.9509202453599], [184.93803519785203, 330.32208588953216], [184.56684569878587, 326.69325153370426], [184.2012860561304, 323.0644171778766], [183.8279010585384, 319.4355828220487], [183.4453340312645, 315.8067484662209], [183.0683599576902, 312.17791411039315], [182.71184748639774, 308.54907975456535], [182.3689224387309, 304.9202453987375], [182.00483902321398, 301.2914110429097], [181.61573833142677, 297.6625766870819], [181.20273685059365, 294.0337423312541], [180.78235795415725, 290.4049079754263], [180.35390785551618, 286.7760736195985], [179.9255940771747, 283.14723926377076], [179.5060134222481, 279.5184049079429], [179.08548331936316, 275.8895705521151], [178.66444396948737, 272.2607361962873], [178.2383960645053, 268.63190184045953], [177.80968742107578, 265.0030674846317], [177.390851716288, 261.37423312880395], [176.9834722927048, 257.74539877297616], [176.58457842563146, 254.11656441714837], [176.196867608581, 250.4877300613205], [175.81904147233882, 246.85889570549273], [175.4451992221618, 243.23006134966496], [175.06507255058798, 239.6012269938372], [174.6805524935781, 235.97239263800932], [174.2905184819413, 232.34355828218153], [173.8953731799266, 228.71472392635377], [173.4918281295416, 225.08588957052592], [173.0970311929409, 221.45705521469813], [172.71229124823054, 217.8282208588703], [172.35391542886708, 214.19938650304255], [172.02034990403462, 210.57055214721473], [171.70452776547907, 206.94171779138696], [171.36265740233216, 203.31288343555914], [170.98595810587554, 199.68404907973132], [170.58404330592387, 196.0552147239036], [170.19575733984172, 192.42638036807577], [169.8288935321587, 188.79754601224798], [169.46977434333263, 185.16871165642016], [169.11728322534364, 181.53987730059234], [168.76947070063474, 177.91104294476455], [168.42199890719465, 174.28220858893678], [168.06462825897015, 170.65337423310896], [167.70096026601766, 167.02453987728117], [167.32813289684805, 163.39570552145338], [166.93387334674952, 159.7668711656256], [166.51830864303503, 156.1380368097978], [166.10414274516765, 152.50920245397], [165.69393986322777, 148.88036809814218], [165.28926363343507, 145.25153374231442], [164.88933670468234, 141.6226993864866], [164.49257145847514, 137.9938650306588], [164.08796890672403, 134.365030674831], [163.65847215721982, 130.7361963190032], [163.20450744249962, 127.10736196317539], [162.72854532481318, 123.4785276073476], [162.24710996794096, 119.84969325151981], [161.79731641468226, 116.22085889569202], [161.33889732382391, 112.59202453987722], [160.88462755528062, 108.96319018404898], [160.4356939811431, 105.33435582822075], [159.99328347348325, 101.70552147239252], [159.55858290437303, 98.07668711656429], [159.1327791458844, 94.44785276073605], [158.7170590700893, 90.81901840490784], [158.31260954905963, 87.1901840490796], [157.9206174548674, 83.56134969325137], [157.54226965958446, 79.93251533742314], [157.17875303528282, 76.30368098159494], [156.83125445403442, 72.67484662576669], [156.50096078791114, 69.04601226993847], [156.18905890898498, 65.41717791411025], [155.89673568932787, 61.78834355828204], [155.62517800101173, 58.15950920245382], [155.37557271610848, 54.53067484662557], [155.14910670669013, 50.901840490797355], [154.94696684482855, 47.27300613496914], [154.77034000259573, 43.64417177914095], [154.62041305206358, 40.01533742331273], [154.49837286530405, 36.38650306748451], [154.4054063143891, 32.757668711656294], [154.3427002713906, 29.128834355828104], [154.3114416083806, 25.499999999999886]])
-CROSSING_LANE_0 = np.array([[406.6965103362119, 788.4999999999997], [406.4832018081058, 783.7975460122697], [406.2675562553765, 779.0950920245397], [406.0496092892277, 774.3926380368097], [405.82939652086304, 769.6901840490796], [405.60695356148614, 764.9877300613496], [405.3823160223007, 760.2852760736196], [405.15551951451033, 755.5828220858895], [404.9265996493188, 750.8803680981595], [404.69559203792966, 746.1779141104295], [404.4625322915466, 741.4754601226994], [404.2274560213733, 736.7730061349694], [403.99039883861343, 732.0705521472393], [403.7513963544706, 727.3680981595091], [403.5104841801485, 722.6656441717791], [403.2676979268508, 717.963190184049], [403.02307320578115, 713.260736196319], [402.7766456281432, 708.5582822085889], [402.52845080514066, 703.8558282208588], [402.2785243479771, 699.1533742331287], [402.0269018678562, 694.4509202453987], [401.7736189759817, 689.7484662576686], [401.5187112835572, 685.0460122699385], [401.26221440178637, 680.3435582822085], [401.00416394187283, 675.6411042944784], [400.7445955149738, 670.9386503066704], [400.4759494687558, 666.2361963189408], [400.2143719722636, 661.5337423312113], [399.9560978112275, 656.8312883434819], [399.69679222419353, 652.1288343557524], [399.43645683526313, 647.4263803680227], [399.16903408482256, 642.7239263802934], [398.88948674846705, 638.0214723925639], [398.6140615291838, 633.3190184048342], [398.3385090220368, 628.6165644171047], [398.0583883518218, 623.9141104293752], [397.7837905297089, 619.2116564416458], [397.516208052801, 614.5092024539161], [397.25361669145036, 609.8067484661866], [396.9678394409739, 605.1042944784572], [396.68211450914447, 600.4018404907275], [396.4087121125336, 595.6993865029981], [396.1260206895123, 590.9969325152686], [395.8476978387838, 586.294478527539], [395.575150091244, 581.5920245398095], [395.2927927710733, 576.88957055208], [395.0074301510861, 572.1871165643506], [394.7221113036723, 567.4846625766211], [394.4375157459214, 562.7822085888914], [394.1543229949236, 558.079754601162], [393.880747119962, 553.3773006134325], [393.6110732572465, 548.6748466257029], [393.3394208047387, 543.9723926379734], [393.0661577299752, 539.2699386502441], [392.7958038638079, 534.5674846625145], [392.5326492658187, 529.865030674785], [392.266354437319, 525.1625766870554], [391.99714060350453, 520.4601226993258], [391.7254125564575, 515.7576687115964], [391.4515598025999, 511.05521472386687], [391.17595656269293, 506.3527607361373], [390.8989617718377, 501.65030674840784], [390.6218076627117, 496.94785276067836], [390.3522363963074, 492.24539877294876], [390.0918236028765, 487.5429447852193], [389.8303778548817, 482.8404907974898], [389.5676634665451, 478.1380368097603], [389.30460178973874, 473.43558282203077], [389.04605612236895, 468.7331288343013], [388.7841217717101, 464.0306748465717], [388.52274854375247, 459.32822085884226], [388.2478083627953, 454.6257668711127], [387.9672456295413, 449.9233128833832], [387.6772877127017, 445.2208588956537], [387.3801836106141, 440.51840490792415], [387.0832588110613, 435.8159509201946], [386.7836263651903, 431.1134969324651], [386.48735728283805, 426.4110429447356], [386.1868199577101, 421.7085889570061], [385.8804104116021, 417.0061349692765], [385.56917180109826, 412.3036809815471], [385.2541099081851, 407.6012269938175], [384.93619314025096, 402.8987730060881], [384.6163525300863, 398.19631901835857], [384.3029172369971, 393.493865030629], [383.99859901532255, 388.79141104289954], [383.69409590663605, 384.08895705517], [383.38385467390043, 379.3865030674405], [383.07693536259455, 374.684049079711], [382.77464115344196, 369.9815950919815], [382.47659809491745, 365.27914110425195], [382.18244534989776, 360.57668711652235], [381.8918351956613, 355.8742331287929], [381.6044330238886, 351.1717791410633], [381.3163221533106, 346.4693251533339], [381.03671789350267, 341.7668711656043], [380.75344542257426, 337.0644171778749], [380.4767430151384, 332.3619631901454], [380.2132916078363, 327.6595092024158], [379.94880253868155, 322.95705521468636], [379.68486694939793, 318.2546012269568], [379.41351349830575, 313.5521472392273], [379.14660838286215, 308.8496932514978], [378.88375698116016, 304.1472392637682], [378.62239975277294, 299.44478527603883], [378.3670144843004, 294.74233128830923], [378.104957719819, 290.03987730057963], [377.8409753355701, 285.3374233128502], [377.56889023112666, 280.63496932512066], [377.30906025734987, 275.9325153373912], [377.0707894570748, 271.2300613496616], [376.8403240448421, 266.5276073619321], [376.61765946100877, 261.82515337420256], [376.39969556518304, 257.12269938647313], [376.17958771651064, 252.4202453987436], [375.9536625623304, 247.71779141101402], [375.72486926239674, 243.01533742328454], [375.4858953651634, 238.31288343555508], [375.2498114999702, 233.61042944782554], [375.0087450652935, 228.90797546009608], [374.7620983223274, 224.2055214723665], [374.49835602716877, 219.503067484637], [374.2294419677379, 214.8006134969075], [373.9508437475724, 210.09815950917798], [373.6849832767491, 205.39570552144846], [373.4285171644332, 200.69325153371895], [373.16137719697497, 195.99079754598944], [372.87973176301716, 191.28834355825992], [372.581409632915, 186.5858895705304], [372.2812331437524, 181.8834355828009], [371.9990567531128, 177.1809815950714], [371.7241014242977, 172.47852760734187], [371.4407439674129, 167.77607361961233], [371.1610008744355, 163.07361963188285], [370.891392689687, 158.3711656441533], [370.628736840219, 153.66871165642382], [370.3648792425441, 148.9662576686943], [370.1118319419014, 144.2638036809648], [369.87979151605407, 139.56134969323526], [369.61758762726663, 134.8588957055214], [369.3570685165318, 130.15644171779132], [369.0984112662082, 125.45398773006127], [368.8417929586117, 120.75153374233119], [368.587390676058, 116.04907975460111], [368.33538150086315, 111.34662576687103], [368.0859425153429, 106.64417177914098], [367.83925080181297, 101.94171779141092], [367.59548344258934, 97.23926380368084], [367.3548175199878, 92.53680981595076], [367.1174301163242, 87.83435582822071], [366.8834983139144, 83.13190184049063], [366.65319919507414, 78.42944785276057], [366.42670984211935, 73.72699386503052], [366.2042073373658, 69.02453987730047], [365.9858687631294, 64.32208588957039], [365.7718712017259, 59.619631901840364], [365.56239173547124, 54.91717791411031], [365.35760744668113, 50.21472392638026], [365.1576954176715, 45.51226993865018], [364.9628327307582, 40.809815950920154], [364.7731964682569, 36.1073619631901], [364.5889637124837, 31.404907975460105], [364.4103115457542, 26.702453987730024], [364.23741705038435, 22.00000000000003]])
-STOPLINE = np.array([[390.86005782319455, 791.9999999999995], [390.9764649300736, 790.2314487632505], [391.0880047548123, 788.4628975265014], [391.19483729634607, 786.6943462897523], [391.29712255361034, 784.9257950530032], [391.3950205255405, 783.1572438162541], [391.4886912110719, 781.388692579505], [391.57829460914, 779.6201413427559], [391.6639907186803, 777.8515901060068], [391.7459395386281, 776.0830388692577], [391.8243010679188, 774.3144876325086], [391.89923530548776, 772.5459363957596], [391.97090225027057, 770.7773851590105], [392.03946190120246, 769.0088339222613], [392.105074257219, 767.2402826855122], [392.1678993172554, 765.4717314487631], [392.2280970802472, 763.7031802120139], [392.2858275451298, 761.9346289752648], [392.34125071083855, 760.1660777385157], [392.3945265763089, 758.3975265017665], [392.44581514047627, 756.6289752650174], [392.49527640227603, 754.8604240282682], [392.54307036064364, 753.0918727915191], [392.58935701451446, 751.32332155477], [392.63429636282393, 749.5547703180208], [392.6780484044617, 747.786219081185], [392.72920535634177, 746.0176678444362], [392.7768679976367, 744.2491166076873], [392.8233155919421, 742.4805653709383], [392.8670762524167, 740.7120141341894], [392.9069928791114, 738.9434628974404], [392.94878626157544, 737.1749116606915], [392.9898717136908, 735.4063604239427], [393.03092875254737, 733.6378091871936], [393.0725223617891, 731.8692579504448], [393.11206537414034, 730.1007067136961], [393.15197195614786, 728.3321554769472], [393.1914493378732, 726.5636042401982], [393.22987654954727, 724.7950530034492], [393.26963378186224, 723.0265017667002], [393.30825088214584, 721.2579505299514], [393.3464073674882, 719.4893992932024], [393.3879896914639, 717.7208480564535], [393.4401680399684, 715.9522968197048], [393.50755672622495, 714.1837455829557], [393.59176410075423, 712.4151943462068], [393.6881856148895, 710.6466431094578], [393.7872138454229, 708.878091872709], [393.8876129156451, 707.1095406359602], [393.98811793530746, 705.340989399211], [394.08743500062275, 703.5724381624624], [394.18334701700616, 701.8038869257133], [394.2722411185497, 700.0353356889643], [394.35342488214116, 698.2667844522157], [394.4276367923869, 696.4982332154666], [394.4950800240965, 694.7296819787177], [394.55857713438814, 692.9611307419689], [394.618001264064, 691.1925795052199], [394.6770404541116, 689.4240282684709], [394.7392788170213, 687.655477031722], [394.8045898937821, 685.8869257949731], [394.8728760388288, 684.1183745582243], [394.9420068180156, 682.3498233214754], [395.0125294066763, 680.5812720847265], [395.08231725247674, 678.8127208479775], [395.1520498725066, 677.0441696112287], [395.22183411662627, 675.2756183744797], [395.289832607442, 673.5070671377308], [395.35800137771696, 671.738515900982], [395.42554765751294, 669.969964664233], [395.49167867689175, 668.2014134274841], [395.55447217885154, 666.4328621907352], [395.6111587910923, 664.6643109539862], [395.6616322640337, 662.8957597172373], [395.70988113888865, 661.1272084804883], [395.761748382219, 659.3586572437393], [395.8205075616025, 657.5901060069906], [395.8868552020654, 655.8215547702417], [395.957078250949, 654.0530035334928], [396.0262145527128, 652.284452296744], [396.0917217209778, 650.5159010599949], [396.1536059586384, 648.747349823246], [396.20968188590916, 646.978798586497], [396.2647763551923, 645.2102473497482], [396.3220556039809, 643.4416961129994], [396.3821657337719, 641.6731448762503], [396.44809365837864, 639.9045936395015], [396.5166402049391, 638.1360424027525], [396.5870126035152, 636.3674911660036], [396.6584180841686, 634.5989399292548], [396.7309801445284, 632.8303886925057], [396.80586056672865, 631.061837455757], [396.8813336654954, 629.293286219008], [396.9580789579187, 627.5247349822592], [397.0358024267982, 625.7561837455102], [397.11332276088694, 623.9876325087612], [397.1921951258873, 622.2190812720124], [397.2716267518613, 620.4505300352635], [397.3508248688705, 618.6819787985146], [397.42899670697716, 616.9134275617657], [397.50534949624273, 615.1448763250166], [397.57833747535324, 613.3763250882678], [397.64577734549425, 611.6077738515189], [397.70924924402095, 609.8392226147699], [397.76786999881, 608.0706713780211], [397.82555279591037, 606.3021201412723], [397.887455842348, 604.5335689045232], [397.95334050676774, 602.7650176677744], [398.02528047683535, 600.9964664310253], [398.10131231633306, 599.2279151942765], [398.17880783884357, 597.4593639575277], [398.25697427442907, 595.6908127207786], [398.3350188531513, 593.9222614840298], [398.4121488050721, 592.1537102472809], [398.48883122815795, 590.3851590105321], [398.5656655621315, 588.6166077737831], [398.6412697220568, 586.848056537034], [398.7163232250233, 585.0795053002853], [398.7901884534936, 583.3109540635362], [398.86224643863176, 581.5424028267873], [398.93370874580273, 579.7738515900386], [399.00378260506847, 578.0053003532895], [399.07304964784123, 576.2367491165408], [399.14192186637325, 574.4681978797919], [399.20913170917424, 572.6996466430428], [399.2731938431285, 570.9310954062939], [399.3325835167258, 569.1625441695451], [399.3875567294257, 567.3939929327961], [399.4398692205266, 565.6254416960472], [399.49175827403445, 563.8568904592983], [399.54500940314136, 562.0883392225495], [399.60156931625926, 560.3197879858005], [399.66129534681414, 558.5512367490516], [399.7276810450364, 556.7826855123027], [399.8002778816241, 555.0141342755537], [399.87701112842643, 553.2455830388049], [399.95790559702124, 551.477031802056], [400.039120775272, 549.708480565307], [400.1197888982449, 547.9399293285582], [400.1973009084494, 546.1713780918092], [400.2699674175492, 544.4028268550603], [400.3346825362121, 542.6342756183114], [400.39344351643587, 540.8657243815625], [400.4515625971237, 539.0971731448135], [400.5113230037199, 537.3286219080645], [400.57668822574703, 535.5600706713158], [400.64591288874385, 533.7915194345669], [400.7142510180485, 532.0229681978179], [400.78001370554887, 530.2544169610691], [400.8430870981168, 528.4858657243202], [400.9033289293649, 526.7173144875712], [400.96263924593114, 524.9487632508223], [401.02257449411525, 523.1802120140733], [401.08560526679514, 521.4116607773243], [401.15342803561185, 519.6431095405756], [401.22639052282693, 517.8745583038267], [401.3025361754343, 516.1060070670777], [401.37755202082826, 514.3374558303288], [401.45092810104234, 512.56890459358], [401.5236541979482, 510.80035335683095], [401.59790608684966, 509.03180212008203], [401.6758128812762, 507.2632508833332], [401.75646855844127, 505.49469964658425], [401.8389670955587, 503.7261484098353], [401.9224024698418, 501.9575971730865], [402.005868658504, 500.1890459363375], [402.09046397406206, 498.4204946995886], [402.17545619483195, 496.6519434628397], [402.2600944504276, 494.8833922260908], [402.3449450050904, 493.1148409893418], [402.428512521036, 491.346289752593], [402.51128318513867, 489.5777385158441], [402.5920107358429, 487.80918727909517], [402.6683341514016, 486.0406360423462], [402.74215860084263, 484.2720848055974], [402.81322792463186, 482.50353356884847], [402.8860101677378, 480.7349823320995], [402.96482490584714, 478.9664310953506], [403.04873097977753, 477.19787985860165], [403.1390995493668, 475.42932862185273], [403.2339756623562, 473.6607773851038], [403.3314649947782, 471.89222614835495], [403.43066152384654, 470.1236749116061], [403.53065922677456, 468.35512367485705], [403.63055208077606, 466.58657243810825], [403.7294340630642, 464.81802120135933], [403.8263991508528, 463.04946996461047], [403.9180470374225, 461.2809187278614], [404.0006028812815, 459.5123674911126], [404.07351790634885, 457.74381625436365], [404.139564846493, 455.9752650176148], [404.20380542371595, 454.2067137808658], [404.2741462075392, 452.43816254411695], [404.3520404361021, 450.66961130736803], [404.4355665327198, 448.9010600706191], [404.5231592069653, 447.13250883387025], [404.61058084139677, 445.36395759712127], [404.6968783512667, 443.59540636037235], [404.78360818287535, 441.8268551236235], [404.8692883246567, 440.0583038868745], [404.9536260400082, 438.28975265012565], [405.0371875931713, 436.52120141337673], [405.1180195125782, 434.7526501766278], [405.19735777498494, 432.98409893987895], [405.27509809172636, 431.21554770312997], [405.3503344400159, 429.44699646638105], [405.42473779959965, 427.67844522963225], [405.4971110118354, 425.90989399288327], [405.5677554173408, 424.14134275613435], [405.6387095673854, 422.3727915193855], [405.7093777439832, 420.6042402826366], [405.7821323341852, 418.8356890458877], [405.857696235324, 417.0671378091387], [405.9351634246133, 415.2985865723898], [406.01362787926627, 413.53003533564083], [406.09218357649684, 411.761484098892], [406.16992449351835, 409.9929328621431], [406.2459446075444, 408.22438162539413], [406.3220866984897, 406.4558303886452], [406.39640013611216, 404.6872791518964], [406.468878517424, 402.9187279151475], [406.5400881066667, 401.1501766783985], [406.60778909865775, 399.3816254416496], [406.6714446428809, 397.6130742049006], [406.7315320820835, 395.8445229681518], [406.78928899376865, 394.07597173140283], [406.8487878783282, 392.30742049465397], [406.91814472300615, 390.5388692579051], [406.9980677783007, 388.77031802115613], [407.08766822945523, 387.00176678440727], [407.1831934853607, 385.23321554765835], [407.27615826644603, 383.4646643109094], [407.365666994799, 381.6961130741605], [407.45159741336954, 379.9275618374116], [407.5337988518488, 378.15901060066267], [407.6131746517721, 376.3904593639138], [407.6881623240194, 374.62190812716483], [407.7585449673761, 372.85335689041597], [407.8271120034978, 371.08480565366705], [407.90269799494496, 369.31625441691807], [407.98880869764315, 367.5477031801692], [408.0853254562234, 365.7791519434203], [408.189123252427, 364.01060070667137], [408.28728297768123, 362.24204946992245], [408.3774052323301, 360.47349823317353], [408.46016493131606, 358.7049469964247], [408.536179762877, 356.9363957596758], [408.611892933671, 355.1678445229268], [408.68742922192456, 353.39929328617797], [408.76188260485117, 351.630742049429], [408.8375539961492, 349.8621908126801], [408.91174297569455, 348.0936395759312], [408.98490127428306, 346.3250883391823], [409.057595156156, 344.55653710243337], [409.12712668234747, 342.78798586568445], [409.1895027495154, 341.01943462893564], [409.2449354566339, 339.25088339218667], [409.2938660095871, 337.48233215543775], [409.34040072454656, 335.7137809186888], [409.401462836464, 333.9452296819786], [409.46347185359963, 332.1766784452295], [409.5266721269649, 330.4081272084804], [409.591308007524, 328.63957597173123], [409.657623846241, 326.87102473498214], [409.72586399408, 325.102473498233], [409.79627280200503, 323.33392226148385], [409.8690946209804, 321.56537102473476], [409.94457380196997, 319.7968197879856], [410.022954695938, 318.02826855123647], [410.10448165384855, 316.2597173144874], [410.1893990266658, 314.4911660777382], [410.2779511653537, 312.7226148409891], [410.37038242087647, 310.95406360424], [410.4669371441982, 309.18551236749084], [410.567859686283, 307.4169611307417], [410.6733943980949, 305.6484098939926], [410.7837856305981, 303.8798586572435], [410.89927773475665, 302.11130742049437], [411.0201150615347, 300.3427561837453], [411.14654196189633, 298.57420494699613], [411.2788027868056, 296.80565371024704], [411.4171418872267, 295.03710247349795], [411.5618036141237, 293.2685512367488], [411.71303231846065, 291.4999999999997]])
